refactor(none): ✨ move unified plugins to under lib/unified
✅ Closes: #80
