+++ src/components/ui/BulletinBoard/BulletinBoard.astro
