+++ src/components/ui/BulletinBoard/index.ts
