+++ src/components/ui/PhotoGallery/PhotoGalleryItem.astro
