+++ src/layouts/BlogLayout.astro
@@ -3,9 +3,9 @@ import 'katex/dist/katex.min.css';
 import type { CollectionEntry } from 'astro:content';
 import { getEntry } from 'astro:content';
 import { Article } from '@/components/models/Article';
-import { BuyMeACoffee } from '@/components/models/BuyMeACoffee';
+import { BuyMeACoffee } from '@/components/ui/BuyMeACoffee';
 import { FormattedDate } from '@/components/models/FormattedDate';
-import { Likes } from '@/components/models/Likes/Likes';
+import { Likes } from '@/components/ui/Likes/Likes';
 import { Taxonomy } from '@/components/models/Taxonomy';
 import { Toc } from '@/components/models/Toc';
 import { Icon } from '@/components/ui/Icon';
