+++ src/layouts/NewsLayout.astro
@@ -1,7 +1,7 @@
 ---
 import type { CollectionEntry } from 'astro:content';
 import { Article } from '@/components/models/Article';
-import { BuyMeACoffee } from '@/components/models/BuyMeACoffee';
+import { BuyMeACoffee } from '@/components/ui/BuyMeACoffee';
 import { FormattedDate } from '@/components/models/FormattedDate';
 import { Icon } from '@/components/ui/Icon';
 import BaseLayout from '@/layouts/BaseLayout.astro';
