+++ src/lib/unified/plugins/rehype-image-figure.ts
