+++ src/pages/[locale]/index.astro
@@ -1,7 +1,7 @@
 ---
 import { getEntry } from 'astro:content';
 import { BlogList } from '@/components/models/BlogList';
-import { BulletinBoard } from '@/components/models/BulletinBoard';
+import { BulletinBoard } from '@/components/ui/BulletinBoard';
 import { PhotoGallery } from '@/components/models/PhotoGallery';
 import { Icon } from '@/components/ui/Icon';
 import BaseLayout from '@/layouts/BaseLayout.astro';
